Fixes for windows (A window* was null in CaveLevel)

+++ Informatik/Informatik/graphics/overlay/LightOverlay.cpp
@@ -60,7 +60,7 @@ void LightOverlay::startFrame()
     // Reset data
     for(int i = 0; i < MAX_LIGHTS; i++)
     {
-        sources[i].x = 0x414570A3; // Some weird stuff I think
+        sources[i].x = (float) 0x414570A3; // Some weird stuff I think
         sources[i].y = -100;
     }
     count = 0;
